move app.js into components directory
